requirements.txt
dydx-v3-python
python-dotenv
pandas
numpy
